refactor : 타입 관련 파일 통합
