--- src/shared/types/profile.ts
@@ -1,48 +1,0 @@
-import { UseFormRegister, FieldValues, Path } from 'react-hook-form';
-import React from 'react';
-
-export type Gender = 'MALE' | 'FEMALE';
-
-export interface GenderOption {
-    value: Gender;
-    label: string;
-    icon?: string;
-}
-
-export interface PetInfo {
-    type: string;
-    age: number;
-    gender: Gender;
-}
-
-export interface ProfileInfo {
-    nickname: string;
-    breed: string;
-    address: string;
-}
-
-export interface ProfileData extends ProfileInfo, PetInfo {
-    profileImage: string | File;
-}
-
-export interface BaseInputProps {
-    label?: string;
-    placeholder?: string;
-    maxLength?: number;
-    type?: string;
-    min?: number;
-    errorMessage?: string;
-    disabled?: boolean;
-    required?: boolean;
-    rightElement?: React.ReactNode;
-}
-
-export interface FormInputProps<T extends FieldValues = FieldValues> extends BaseInputProps {
-    name: Path<T>;
-    register: UseFormRegister<T>;
-}
-
-export interface FormFieldProps<T extends FieldValues = FieldValues> extends BaseInputProps {
-    name: Path<T>;
-    register: UseFormRegister<T>;
-}
